requirements.txt
Django==1.8
djangorestframework==3.3.2
